requirements-dev.txt
amqp==2.5.2
asgi-redis==1.4.3
asgiref==1.1.2
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
billiard==3.6.3.0
bleach==3.1.1
boto3==1.12.21
botocore==1.15.21
celery==4.4.1
certifi==2019.11.28
cffi==1.14.0
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
daphne==1.4.2
Django==2.2.11
django-channels==0.7.0
django-channels-presence==0.0.9
django-cors-headers==3.2.1
django-rest-swagger==2.2.0
django-storages==1.9.1
django-webpack-loader==0.7.0
djangorestframework==3.11.0
docutils==0.15.2
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.5.0
incremental==17.5.0
itypes==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
kombu==4.6.8
Markdown==3.2.1
MarkupSafe==1.1.1
msgpack-python==0.5.6
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
redis==2.10.6
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
service-identity==18.1.0
simplejson==3.17.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
Twisted==19.10.0
txaio==20.1.1
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
webencodings==0.5.1
whitenoise==3.3.1
zipp==3.1.0
zope.interface==4.7.2
